created edly app in openedx features to bring dynamic theming logic at one place
