+++ lms/envs/devstack_docker.py
@@ -71,6 +71,7 @@ MKTG_URLS = {
     'TOS_AND_HONOR': '/edx-terms-service',
     'WHAT_IS_VERIFIED_CERT': '/verified-certificate',
     "NAV_MENU": "wp-json/edly-wp-routes/nav-menu",
+    "FOOTER": "wp-json/edly-wp-routes/footer",
     "ZENDESK-WIDGET": "wp-json/edly-wp-routes/edly-zendesk-widget"
 }
