+++ openedx/features/edly/apps.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from django.apps import AppConfig
+
+
+class EdlyAppConfig(AppConfig):
+    name = 'edly_app'
